+++ smash-nba/src/app/StyleComponents/Carrito.css
@@ -57,8 +57,8 @@
   
   .delete-button {
     position: absolute;
-    top: 10px;
-    right: 10px;
+    top: -30px;
+    left: 0px;
   }
