+++ smash-nba/src/app/StyleComponents/CartaContainer.scss
@@ -46,6 +46,7 @@ $color3: #725b16;
 	}
 	#card-inner{ 
 		#popup-text{
+
 			background: linear-gradient(135deg, #BDCAD9 0%, darken( #4847a9, 10%) 100%);
 		}
 		#card-top{
@@ -155,7 +156,6 @@ $color3: #725b16;
 		filter: brightness(120%); 
 	  }
 }
-#body:hover{filter: brightness(120%); }
 #card {
 	position: relative;
 	color: $color3;
@@ -174,16 +174,34 @@ $color3: #725b16;
 		background: #aaa;
 		height: 100%;
 		width: 100%;
-		#popup-text{
+		#popup-text {
 			position: absolute;
 			top: 0;
 			left: 0;
 			width: 100%;
 			height: 54%;
 			overflow: hidden;
-			padding-top:50%;
+			display: flex;
+			align-items: center; /* Center vertically */
+			justify-content: center; /* Center horizontally */
 			
-		}
+		  }
+		  
+		  .popup-content {
+			display: flex;
+			flex-direction: column;
+			align-items: center; /* Center horizontally within the flex container */
+			justify-content: center; /* Center vertically within the flex container */
+		  }
+		  
+		  .popup-content p {
+			margin-bottom: 10px; /* Add spacing between the elements */
+		  }
+		  
+		  .popup-content h1 {
+			margin: 0; /* Remove any default margin */
+		  }
+		  
 		#card-top {
 			position: absolute;
 			top: 0;
