+++ smash-nba/src/app/StyleComponents/Inicio.css
@@ -1,0 +1,3 @@
+.container{
+    width:100%;
+}
